requirements.txt
opencv-python==4.12.0.88
transformers==4.56.2
torch
pillow
torchvision
accelerate
pandas
matplotlib